requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pandas>=1.5.0
pyyaml>=6.0
notion-client==2.2.1
python-dotenv==1.0.0
schedule>=1.2.0